chore: little refactors
